main/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-multipart==0.0.20
aiofiles==24.1.0
httpx==0.28.1
pyjwt==2.8.0
